#21: Making some in-roads. To make our PHP binding downcasting aware, we need to compile a static class id -> name map (just like for .net) and use that to resolve the appropriate swig_type_info through the mapped class name. This means that we now properly get proper derived types in APIs like MgSiteConnection.CreateService, but we now have the problem of some of the generated derived classes (eg. MgResourceService) being abstract, which is true. They are indeed abstract in C++, but in PHP we actually don't care.
The underlying type (which in this example would be MgProxyResourceService) is upcastable to MgResourceService, but MgProxyResourceService (and similar classes) are not exposed to the public API surface. So how can we address this? I think we just find a way to instruct SWIG to not bother generating abstract classes. How? I have no idea right now.