#21: Add a new PhpPostProcess tool to manually fix up inheritance chains (mainly on MgService subclasses) after the initial MapGuideApi.php is generated.
With this, the remaining major problem is with overloads, which I suspect is mainly due to stricter types in PHP7. For example "-1" doesn't appear to be an acceptable value to pass to an int parameter any more. Fix some parts of the PHP test suite where this is the case. More to follow.